src/frontend/requirements.txt
click==8.1.3
Flask==2.2.2
fluent-logger==0.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
lorem-text==2.1
MarkupSafe==2.1.2
msgpack==1.0.4
waitress==2.1.2
Werkzeug==2.2.2